refactor(cli): extract the playground session from the xterm hook
The hook only mounts and fits the terminal. Typing, history, the pager, and the attract loop live in one session so each file has a single job.
