+++ src/components/pages/cli/playground.js
@@ -2,8 +2,9 @@ import React, { useRef } from 'react'
 
 import Terminal from 'components/elements/Terminal/Terminal'
 
-import { theme } from 'theme'
+import { theme, toPx } from 'theme'
 
+import { XTERM_SURFACE_CSS } from './shared'
 import { PLAYGROUND_HEIGHT, useCliTerminal } from './use-cli-terminal'
 
 import '@xterm/xterm/css/xterm.css'
@@ -11,7 +12,6 @@ import '@xterm/xterm/css/xterm.css'
 const Playground = () => {
   const surfaceRef = useRef(null)
   useCliTerminal(surfaceRef)
-  const height = PLAYGROUND_HEIGHT
 
   return (
     <Terminal
@@ -31,9 +31,6 @@ const Playground = () => {
         },
         '& > div:first-of-type *': {
           color: 'white70'
-        },
-        '& .xterm-rows > div > span': {
-          padding: 0
         }
       })}
       aria-label='Interactive Microlink CLI'
@@ -42,18 +39,11 @@ const Playground = () => {
         ref={surfaceRef}
         css={theme({
           width: '100%',
-          height: `${height}px`,
-          minHeight: `${height}px`,
+          height: toPx(PLAYGROUND_HEIGHT),
+          minHeight: toPx(PLAYGROUND_HEIGHT),
           overflow: 'hidden',
           bg: 'black',
-          '& .xterm': { height: '100%' },
-          '& .xterm span': {
-            padding: 0
-          },
-          '& .xterm-viewport': {
-            overflowY: 'auto',
-            bg: 'black'
-          }
+          ...XTERM_SURFACE_CSS
         })}
       />
     </Terminal>
